Restructure codebase in to multiple files. Add NPM to manage dependences. Setup rollup to build/watch Updated HACs file to reference built copy in dist.
Initially organised around classes, but util likely makes more sense to separate into discrete methods at some point.
Some of the class types are a little arbitrary, but hopefully this arrangement makes rough sense.
