+++ hacs.json
@@ -1,5 +1,5 @@
 {
     "name": "map-card",
     "render_readme": true,
-    "filename":"map-card.js"
+    "filename":"dist/map-card.js"
 }
