+++ package.json
@@ -1,0 +1,29 @@
+{
+  "name": "ha-map-card",
+  "version": "0.12.0",
+  "description": "A better Map Card for Home Assistant",
+  "main": "dist/map-card.js",
+  "scripts": {
+    "build": "rollup -c --bundleConfigAsCjs",
+    "watch": "rollup -c --watch --bundleConfigAsCjs"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/nathan-gs/ha-map-card.git"
+  },
+  "author": "Nathan Bijnens",
+  "license": "MIT",
+  "bugs": {
+    "url": "https://github.com/nathan-gs/ha-map-card/issues"
+  },
+  "homepage": "https://github.com/nathan-gs/ha-map-card#readme",
+  "dependencies": {
+    "leaflet": "^1.9.4",
+    "lit": "^3.1.3"
+  },
+  "devDependencies": {
+    "@rollup/plugin-commonjs": "^25.0.8",
+    "@rollup/plugin-node-resolve": "^15.2.3",
+    "rollup": "^4.18.0"
+  }
+}
