+++ src/configs/LayerConfig.js
@@ -1,0 +1,12 @@
+export default class LayerConfig {
+  /** @type {String} */
+  url;
+  /** @type {Object} */
+  options;
+
+  constructor(url, options, attribution = null) {
+    this.url = url;
+    this.options = {...{attribution: attribution}, ...options};
+  }
+
+}
