+++ src/configs/TileLayerConfig.js
@@ -1,0 +1,3 @@
+import LayerConfig from "./LayerConfig.js";
+
+export default class TileLayerConfig extends LayerConfig {};
