+++ src/configs/WmsLayerConfig.js
@@ -1,0 +1,3 @@
+import LayerConfig from "./LayerConfig.js";
+
+export default class WmsLayerConfig extends LayerConfig {};
