+++ src/models/TimelineEntry.js
@@ -1,0 +1,14 @@
+export default class TimelineEntry {  
+  /** @type {Date} */
+  timestamp;
+  /** @type {Double} */
+  latitude;
+  /** @type {Double} */
+  longitude;
+
+  constructor(timestamp, latitude, longitude) {  
+    this.timestamp = timestamp;  
+    this.latitude = latitude;  
+    this.longitude = longitude;  
+  }  
+}
